+++ shared/lib/emojis.ts
@@ -1,3 +1,7 @@
+// SPDX-FileCopyrightText: 2024 John Livingston <https://www.john-livingston.fr/>
+//
+// SPDX-License-Identifier: AGPL-3.0-only
+
 export const maxSize: number = 30 * 1024
 export const allowedExtensions = ['png', 'jpg', 'jpeg', 'gif']
 export const inputFileAccept = ['image/jpg', 'image/png', 'image/gif']